オプションの組み合わせに対応
